+++ gulpfile.js
@@ -3,8 +3,8 @@ const gulp = require('gulp');
 const $ = require('gulp-load-plugins')();
 
 // Import configurations
-const paths = require('./gulpfile.paths');
-const _ = require('./gulpfile.helpers');
+const paths = require('./tasks/config/paths');
+const _ = require('./tasks/config/helpers');
 
 let tasks = require('require.all')('./tasks');
 tasks((name, task) => { func = () => task(gulp, paths, $, _, tasks); func.displayName = name; return func});
